% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_plugin.R
\name{inline_plugin}
\alias{inline_plugin}
\title{Create Plugin}
\usage{
inline_plugin(dev, prod)
}
\arguments{
\item{dev, prod}{Development and production
functions to run when in either state. This
callback function should accept a single argument
a vector of character strings: the source code where
each character string is a line of code.}
}
\description{
Create an inline plugin.
}
\examples{
chg_magrittr_to_native <- function(lines){
  gsub("\\\\\%>\\\\\%", "|>", lines)
}

code <- 'function(x){
  x \%>\% 
    sum()
}'

chg_magrittr_to_native(code)
}
