% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_package}
\alias{parse_package}
\title{Rich parse data for all source files in a package}
\usage{
parse_package(path = ".")
}
\description{
Returns a tibble with one row per function/object in the package.
}
