% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpp_dev.R, R/rpp_prod.R
\name{rpp_dev_roclet}
\alias{rpp_dev_roclet}
\alias{rpp_prod_roclet}
\title{Roclets}
\usage{
rpp_dev_roclet()

rpp_prod_roclet()
}
\description{
These functions implement \code{\link[roxygen2:roclet]{roxygen2::roclet()}}s for integration
with the roxygen2 package.
}
