% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpp-package.R
\docType{package}
\name{rpp-package}
\alias{rpp}
\alias{rpp-package}
\title{rpp: R Preprocessor}
\description{
A configurable framework for applying code transformations to R package projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rpp.q-lang.org/}
  \item \url{https://github.com/Q-language/rpp}
  \item Report bugs at \url{https://github.com/Q-language/rpp/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

Other contributors:
\itemize{
  \item John Coene \email{jcoenep@gmail.com} [contributor]
}

}
\keyword{internal}
