% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpp_prod.R
\name{rpp_to_prod}
\alias{rpp_to_prod}
\alias{rpp_prod_roclet}
\title{Production State}
\usage{
rpp_to_prod()

rpp_prod_roclet()
}
\description{
Switch a codebase to its production state.
}
\details{
The production state of the code does
not include type checking or other checks, for
performance reasons. You may switch to the
development state of the code with \code{\link[=rpp_to_dev]{rpp_to_dev()}},
but ensure you switch back to production when
pushing the code to online repositories or to CRAN.
}
