% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpp_dev.R
\name{rpp_to_dev}
\alias{rpp_to_dev}
\title{Development State}
\usage{
rpp_to_dev()
}
\description{
Switch to codebase to the development state.
}
\details{
The development state of the codebase
may include type checking and more but cannot
be submitted to CRAN and impacts performances
of runtime code, ensure you switch back to
the production version with \code{\link[=rpp_to_prod]{rpp_to_prod()}}
after development work is done.
}
